airflow/requirements.txt
yfinance
python-dotenv